pyproject.toml
[tool.poetry]
name = "decalls_stats"
version = "0.1.0"
description = ""
authors = ["John Doe <john.doe.github.com>"]


[tool.poetry.dependencies]
python = "^3.10"
solana = "^0.29.2"
python-dotenv = "^1.0.0"
anchorpy = { extras = ["cli"], version = "^0.16.0" }
matplotlib = "^3.7.1"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
